Moved Node class to seperate file in same directory
