+++ queue/queue.py
@@ -14,20 +14,7 @@ Stretch: What if you could only use instances of your Stack class to implement t
          What would that look like? How many Stacks would you need? Try it!
 """
 
-class Node:
-    def __init__(self, value, next=None):
-        self.value = value
-        self.next = next
-
-    def get_value(self):
-        return self.value
-
-    def get_next(self):
-        return self.next
-
-    def set_next(self, new_next):
-        self.next = new_next
-
+from node import Node
 
 class Queue:
     def __init__(self):
