+++ stack/node.py
@@ -1,0 +1,14 @@
+
+class Node:
+    def __init__(self, value, next=None):
+        self.value = value
+        self.next = next
+
+    def get_value(self):
+        return self.value
+
+    def get_next(self):
+        return self.next
+
+    def set_next(self, new_next):
+        self.next = new_next
