+++ stack/stack.py
@@ -11,19 +11,7 @@ return elements in Last In First Out order.
    implementing a Stack?
 """
 
-class Node:
-    def __init__(self, value, next=None):
-        self.value = value
-        self.next = next
-
-    def get_value(self):
-        return self.value
-
-    def get_next(self):
-        return self.next
-
-    def set_next(self, new_next):
-        self.next = new_next
+from node import Node
 
 # Implemented using a linked list
 class Stack:
